Added basic accuracy and critical hit checks (#87)
* Added the basic check information for a move

* Added the basic layout of the critical calculations

* Added some basic critical hit rate modification traits

* Changed trait utilities to use some template code

* Updated the update code to use a more generic template approach

* Fixed MSVC build issues

* Made the definition hack a windows only convention

* Added a new trait to override the crit rate

* Updated the damage calc test

* Fixed the compiler setup

* Restored old member accessor fix

* Fixed lone test failure

* Began transition over to GoogleMock instead of FakeIt

* Additional code changeover

* Removed all references to FakeIt from the code

* Removed RTTI from a bunch of places

* Removed most instances of member_accessor

* Fixed all test issues from the mocking and private access refactor

* Added a test for the standard critical hit roll

* Added additional tests to increase coverage

+++ Plugins/PokemonStarterKit/Source/PokemonCore/Private/Pokemon/GamePokemon.cpp
@@ -76,7 +76,7 @@ bool UGamePokemon::IsShiny() const {
     uint32 B = A & 0xFFFF;
     uint32 C = (A >> 16) & 0xFFFF;
     uint32 D = B ^ C;
-    return D < Pokemon::FBaseSettings::Get().GetShinyPokemonChance();
+    return D < static_cast<uint32>(Pokemon::FBaseSettings::Get().GetShinyPokemonChance());
 }
 
 int32 UGamePokemon::GetCurrentHP() const {
